r110: removed k8_open and fastx functions
The new k8_open file I/O functions are a little more convenient but not much.
As we need to maintain backward compatability indefinitely, we will keep the
old File class anyway. With k8_open etc, we would have two versions of APIs
doing the same thing and the File APIs are actually more flexible.
Performance-wise, new v8 does an impressive job for fastx parsing with File
APIs. v8 APIs are changing fast. It is safer to have a thinner implementation
for long-term maintanence.